fix data type

+++ localization/ndt_scan_matcher/include/ndt_scan_matcher/util_func.hpp
@@ -66,7 +66,7 @@ Eigen::Affine3d fromRosPoseToEigen(const geometry_msgs::msg::Pose & ros_pose);
 
 std::vector<geometry_msgs::msg::Pose> createRandomPoseArray(
   const geometry_msgs::msg::PoseWithCovarianceStamped & base_pose_with_cov,
-  const size_t particle_num);
+  const int particle_num);
 
 template <class T>
 T transform(const T & input, const geometry_msgs::msg::TransformStamped & transform)
